Installation adjustments
The installation structure has been changed.
This is mainly due to the fact that the old nrfutil installed
via pip no longer works with lastest python versions.
The new nrfutil executable from Nordic is now used instead.

A separate Pyhon environment is now also created to handle the
deprecation of system wide pip installations in the newer Python
versions.

These changes require a rerun of the installation script in
already existing installations. This will be detected in
enrfmake.

The obsolete script changed_defs is removed.